Moving quant_pdf tests into their own test module.
